changed path of current DB
