--- DB/AssignmentIDDAO.java
@@ -1,9 +1,0 @@
-package com.example.graderecorder.DB;
-
-import java.util.List;
-
-@Dao
-public interface AssignmentIDDAO {
-
-
-}
