--- DB/CourseDatabase.java
@@ -1,4 +1,0 @@
-package com.example.graderecorder.DB;
-
-public abstract class CourseDatabase extends RoomDatabase{
-}
