--- DB/StudentIDDAO.java
@@ -1,4 +1,0 @@
-package com.example.graderecorder.DB;
-
-public interface StudentIDDAO {
-}
