+++ app/src/main/java/com/parkererickson/a438_project1/DB/AssignmentIDDAO.java
@@ -1,0 +1,9 @@
+package com.example.graderecorder.DB;
+
+import java.util.List;
+
+@Dao
+public interface AssignmentIDDAO {
+
+
+}
