+++ app/src/main/java/com/parkererickson/a438_project1/DB/GradeIDDAO.java
@@ -1,0 +1,4 @@
+package com.example.graderecorder.DB;
+
+public interface GradeIDDAO {
+}
