+++ app/src/main/java/com/parkererickson/a438_project1/DB/StudentDatabase.java
@@ -1,0 +1,4 @@
+package com.example.graderecorder.DB;
+
+public abstract class StudentDatabase extends RoomDatabase{
+}
